fixed meta data

+++ src/entities/meta/CategoryMeta.ts
@@ -5,7 +5,7 @@ export class CategoryMeta {
   static columns = {
     id: {
       name: 'id',
-      type: 'bigint',
+      type: 'integer',
       isPrimary: true,
       isGenerated: true,
     },
@@ -15,7 +15,7 @@ export class CategoryMeta {
     },
     parentCategoryId: {
       name: 'parent_category_id',
-      type: 'bigint',
+      type: 'integer',
       isNullable: true,
     },
   };
